+++ libs/iso-bmff/src/boxes/PreselectionGroupBox.ts
@@ -2,6 +2,14 @@ import type { Entity } from './Entity.ts'
 import type { FullBox } from './FullBox.ts'
 
 /**
+ * Child boxes of Preselection Group Box
+ *
+ * @public
+ */
+// TODO: Find documentation on what children are allowed
+export type PreselectionGroupBoxChild = any;
+
+/**
  * ISO/IEC 14496-12:202x - 8.18.4.1 Preselection group box
  *
  * @public
@@ -21,4 +29,5 @@ export type PreselectionGroupBox = FullBox & {
 	preselectionTag?: string;
 	selectionPriority?: number;
 	interleavingTag?: string;
+	boxes: PreselectionGroupBoxChild[];
 };