+++ libs/iso-bmff/src/boxes/VisualSampleEntryBox.ts
@@ -2,11 +2,18 @@ import type { SampleEntryBox } from './SampleEntryBox.ts'
 import type { VisualSampleEntryType } from './VisualSampleEntryType.ts'
 
 /**
+ * Child boxes of VisualSampleEntryBox
+ *
+ * @public
+ */
+export type VisualSampleEntryBoxChild = any;
+
+/**
  * ISO/IEC 14496-15:2014 - 12.1.3.1 avc1/2/3/4, hev1, hvc1, encv
  *
  * @public
  */
-export type VisualSampleEntryBox<T extends VisualSampleEntryType> = SampleEntryBox & {
+export type VisualSampleEntryBox<T extends VisualSampleEntryType = VisualSampleEntryType> = SampleEntryBox & {
 	type: T;
 	preDefined1: number;
 	reserved2: number;
@@ -20,5 +27,5 @@ export type VisualSampleEntryBox<T extends VisualSampleEntryType> = SampleEntryB
 	compressorName: number[];
 	depth: number;
 	preDefined3: number;
-	config: Uint8Array;
+	boxes: VisualSampleEntryBoxChild[];
 };