+++ libs/iso-bmff/src/readers/readAvc4.ts
@@ -1,6 +1,6 @@
 import type { VisualSampleEntryBox } from '../boxes/VisualSampleEntryBox.ts'
 import type { IsoBoxReadView } from '../IsoBoxReadView.ts'
-import { readAvc1 } from './readAvc1.ts'
+import { readVisualSampleEntryBox } from './readVisualSampleEntryBox.ts'
 
 /**
  * Parse a VisualSampleEntryBox from an IsoView
@@ -12,8 +12,5 @@ import { readAvc1 } from './readAvc1.ts'
  * @public
  */
 export function readAvc4(view: IsoBoxReadView): VisualSampleEntryBox<'avc4'> {
-	return {
-		...readAvc1(view),
-		type: 'avc4',
-	}
+	return readVisualSampleEntryBox('avc4', view)
 }