+++ libs/iso-bmff/src/writers/writeHvc1.ts
@@ -1,16 +1,18 @@
 import type { VisualSampleEntryBox } from '../boxes/VisualSampleEntryBox.ts'
 import type { IsoBoxWriteView } from '../IsoBoxWriteView.ts'
+import type { IsoBoxWriteViewConfig } from '../IsoBoxWriteViewConfig.ts'
 import { writeVisualSampleEntryBox } from './writeVisualSampleEntryBox.ts'
 
 /**
  * Write a VisualSampleEntryBox (hvc1) to an IsoDataWriter.
  *
  * @param box - The VisualSampleEntryBox fields to write
+ * @param config - The IsoBoxWriteViewConfig to use
  *
  * @returns An IsoDataWriter containing the encoded box
  *
  * @public
  */
-export function writeHvc1(box: VisualSampleEntryBox<'hvc1'>): IsoBoxWriteView {
-	return writeVisualSampleEntryBox(box, 'hvc1')
+export function writeHvc1(box: VisualSampleEntryBox<'hvc1'>, config: IsoBoxWriteViewConfig): IsoBoxWriteView {
+	return writeVisualSampleEntryBox(box, config)
 }