+++ libs/iso-bmff/test/readers/readAvc1.test.ts
@@ -28,6 +28,6 @@ describe('readAvc1', function () {
 			0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00]) // length + 'GoPro AVC encoder'
 		assert.strictEqual(box.depth, 24)
 		assert.strictEqual(box.preDefined3, -1)
-		assert.strictEqual(box.config.byteLength, 98)
+		assert.strictEqual(box.boxes.length, 2)
 	})
 })